Refactor: separate route functions from main
