+++ app/routes/hello.py
@@ -1,0 +1,2 @@
+def hello():
+    return "Serviço de nomes habilitado! Volte para a NoHarm e use o sistema normalmente ;)"
